refactor: prover.ts
